% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emsMultifilter.R
\name{emsMultifilter}
\alias{emsMultifilter}
\title{emsMultifilter}
\usage{
emsMultifilter(qry, filter_vector)
}
\arguments{
\item{qry}{a FltQuery object}

\item{filter_vector}{a string vector of filter statements of the form c("'Dimension' == 'Value',...)}
}
\value{
This function returns a FltQuery object with all of the filters in filter_vector applied to it.
}
\description{
Apply multiple filters in one function call as a vector.
}
\details{
Usually you have to make multiple function calls to apply multiple filters - this simplifies that.
}
\examples{
\dontrun{
filter_vector <- c("'Fleet' == 'A380'", "'Flight Date (Exact)' == '2019-01-01'")
filtered_qry <- emsMultifilter(qry, filter_vector)
}
}
\author{
Matt Simmons mattsimmons@qantas.com.au
}
