% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_basic_basic_flight_query.R
\name{generate_basic_flight_query}
\alias{generate_basic_flight_query}
\title{generate_basic_flight_query}
\usage{
generate_basic_flight_query(
  connection_object,
  ems_name,
  db_name,
  reload_preset_fieldtree = FALSE
)
}
\arguments{
\item{connection_object}{a ems connection object}

\item{ems_name}{a string, the name of the EMS instance.}

\item{db_name}{a string, the name of a .db file to read the tree in. Will attempt to use existing if present.}

\item{reload_preset_fieldtree}{a boolean to force the preset fieldtree to reload anyway}
}
\value{
Returns a fltqry object for flights with the basic fieldtree updated and some filtering set.
}
\description{
Shortcut to create a basic flight query object
}
\details{
Creates a new flight query object with a updated fieldtree, and takeoff/landing valid fields
filtered to true. Designed to shortcut some of the boilerplate required for EMS queries.
}
\examples{
\dontrun{
con <- getEMSCon("joe.smith")
data_file <- "mydb.db"
basic_qry <- generate_basic_flight_query(con, "My-Ems", data_file)
}
}
\author{
Matt Simmons mattsimmons@qantas.com.au
}
