% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_emsapi_token.R
\name{get_emsapi_token}
\alias{get_emsapi_token}
\title{Get EMS API Token}
\usage{
get_emsapi_token(username, password)
}
\arguments{
\item{username}{the EMS username to be used.}

\item{password}{the EMS password to be used}
}
\value{
This function returns a string that can be used as the authorization token for other API calls.
}
\description{
A simple function to get an EMS API token
}
\details{
A simple implementation that gets the token required for other API calls.
}
\examples{
\dontrun{
token <- get_emsapi_token("first.name", "password")
}
}
\author{
Matt Simmons mattsimmons@qantas.com.au
}
