% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_generate_preset_fieldtree.R
\name{optional_generate_preset_fieldtree}
\alias{optional_generate_preset_fieldtree}
\title{optional_generate_preset_fieldtree}
\usage{
optional_generate_preset_fieldtree(qry, data_file = NA, full = FALSE)
}
\arguments{
\item{qry}{a Flt_Qry object}

\item{data_file}{an optional .db file to read the tree in, if present}

\item{full}{a boolean to force the preset fieldtree to reload anyway}
}
\value{
Returns a fltqry object with a fieldtree loaded
}
\description{
wrapper for generate_preset_fieldtree that optionally tries to load from a local db if present.
}
\details{
Ideally you'd only ever have to do this once, but not have to write two kinds of code, this is an attempt at that.
}
\examples{
\dontrun{
con <- getEMSCon("joe.smith")
data_file <- "mydb.db"
qry <- flt_query(conn = con, ems_name = "my-ems", data_file = data_file) \%>\%
Rems::set_database("FDW Flights") \%>\%
optional_generate_preset_fieldtree(data_file = data_file, full = FALSE)
}
}
\author{
Matt Simmons mattsimmons@qantas.com.au
}
