% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_kml_flight_track.R
\name{pull_kml_flight_track}
\alias{pull_kml_flight_track}
\title{Pull KML Flight Track}
\usage{
pull_kml_flight_track(
  username,
  password,
  flight_record,
  trajectory_config = "Flight Pulse 1.0"
)
}
\arguments{
\item{username}{the EMS username to be used.}

\item{password}{the EMS password to be used}

\item{flight_record}{a numeric, indicating the flight record to be used for generation}

\item{trajectory_config}{a string indicating the trajectory config to be used. The available ones can be viewed at https://qfa-api.au.efoqa.com:443/api/v2/ems-systems/1/trajectory-configurations}
}
\value{
This function returns the KML data as text XML.To use it, write it to a file with a .kml extension.
}
\description{
Uses the KML API to retrieve and write a KML trajectory directly
}
\details{
Sometimes you just want a KML file. This does that. Only KML types prebuilt are available -
you'll need to check the API to see which those are.
}
\examples{
\dontrun{
flight_track <- pull_kml_flight_track("firstname.lastname", "password", flight_record = 1)
write(flight_track, "flight-track.kml")
}
}
\author{
Matt Simmons mattsimmons@qantas.com.au
}
