% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qEmsTools.R
\docType{package}
\name{qEmsTools}
\alias{qEmsTools}
\alias{qEmsTools-package}
\title{qEmsTools}
\description{
Qantas Tools for using Rems, the R API interface to GE's Event Management System
}
\details{
This package provides some helper functions for interacting with EMS, including
secure password management and other extensions of the query language.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mattsq/qEmsTools}
  \item Report bugs at \url{https://github.com/mattsq/qEmsTools/issues}
}

}
