% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_multiflts_parallel.R
\name{run_multiflts_parallel}
\alias{run_multiflts_parallel}
\title{run_multiflts_parallel}
\usage{
run_multiflts_parallel(flight_record_vector, ts_query)
}
\arguments{
\item{flight_record_vector}{A vector of flight records}

\item{ts_query}{a time series FltQuery object}
}
\value{
This function returns a list of time_series data lists.
}
\description{
Runs run_multiflts in parallel
}
\details{
The longest pole in the tent with multiflts is usually each query - this one makes multiple calls in parallel, as many as your computer will allow.
}
\examples{
\dontrun{
ts_multidata <- tsDataExtractor(c(7835627, 1351387, 34543534), ts_query = ts_query)
}
}
\author{
Matt Simmons mattsimmons@qantas.com.au
}
