% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_fieldtree_vec.R
\name{update_fieldtree_vec}
\alias{update_fieldtree_vec}
\title{Update Fieldtree (Vectorized)}
\usage{
update_fieldtree_vec(qry, tree, exclude_tree = c())
}
\arguments{
\item{qry}{a Rems fltqry object, to be updated with an extended fieldtree}

\item{tree}{a string vector, where each string is the parent folder name. Partial matching is supported.}

\item{exclude_tree}{whatever exclude_tree does??}
}
\value{
This function returns an updated fltqry object.
}
\description{
Rems::update_fieldtree, but takes tree as vector, instead of arguments
}
\details{
Rems::update_fieldtree takes the tree positions as arguments, which is more intuitive but annoying
when you want to reuse a tree multiple times. This wraps the function and takes a string vector instead.
}
\examples{
\dontrun{
base_tree <- c("profiles", "standard", "p1")
qry2 <- update_fieldtree_vec(qry, c(base_tree, "processing"), exclude_tree = c())
}
}
\author{
Matt Simmons mattsimmons@qantas.com.au
}
