% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEMSCon.R
\name{getEMSCon}
\alias{getEMSCon}
\title{getEMSCon}
\usage{
getEMSCon(username)
}
\arguments{
\item{username}{An EMS username as a string, usually something like "firstname.lastname"}
}
\value{
This function returns a EMS connection object to be used in a FltQuery call.
}
\description{
Get an EMS Connection object using prompts
}
\details{
This securely creates an EMS connection object by using the RStudio prompt feature, so it's not stored in your History.
}
\examples{
\dontrun{
con <- getEMSCon("joe.bloggs")
qry <- Rems::flt_query(conn = con, ems_name = "my-ems", data_file = "example.db")
}
}
\author{
Matt Simmons mattsimmons@qantas.com.au
}
