% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_df_flight_track.R
\name{pull_df_flight_track}
\alias{pull_df_flight_track}
\title{Pull Dataframe Flight Track}
\usage{
pull_df_flight_track(
  username,
  password,
  flight_record,
  start = NULL,
  end = NULL,
  verbose = FALSE
)
}
\arguments{
\item{username}{the EMS username to be used.}

\item{password}{the EMS password to be used}

\item{flight_record}{a numeric, indicating the flight record to be used for generation}

\item{start}{a numeric, indicating the seconds offset the trajectory should start at (optional, defaults to start of record)}

\item{end}{a numeric, indicating the seconds offset the trajectory should end at (optional, defaults to end of record)}

\item{verbose}{a logical, indicating whether to print progress - useful for mapping over multiple flight records or large flights.}
}
\value{
This function returns a data frame with offset, latitude, longitude, altitude and flight record details.
}
\description{
Uses the Trajectory API to retrieve a flight track directly
}
\details{
Sometimes you just want a flight track. This uses the simple trajectories API
to pull a dataframe with just the offset, latitude, longitude and altitude of a flight by second.
Not entirely clear just yet what parameters specifically are used.
}
\examples{
\dontrun{
flight_track <- pull_df_flight_track("firstname.lastname", "password", flight_record = 1)
}
}
\author{
Matt Simmons mattsimmons@qantas.com.au
}
